chore: Rename store
