--- mobile/src/stores/HomeViewPreference/actions/index.ts
@@ -1,1 +1,0 @@
-export * as HomeViewPreferenceSetters from "./setters";
