+++ mobile/src/stores/ViewPreference/actions/index.ts
@@ -1,0 +1,1 @@
+export * as ViewPreferenceSetters from "./setters";
