+++ mobile/src/stores/ViewPreference/actions/setters.ts
@@ -1,7 +1,7 @@
-import { homeViewPreferenceStore } from "../store";
+import { viewPreferenceStore } from "../store";
 import type { LayoutOption } from "../constants";
 import type { MutableLayout } from "../types";
 
 export function setLayout(screen: MutableLayout, layout: LayoutOption) {
-  homeViewPreferenceStore.setState({ [`${screen}Layout`]: layout });
+  viewPreferenceStore.setState({ [`${screen}Layout`]: layout });
 }
