+++ mobile/src/stores/ViewPreference/constants.ts
@@ -5,11 +5,11 @@ export type LayoutOption = (typeof LayoutOptions)[number];
 //#endregion
 
 //#region Store
-export interface HomeViewPreferenceStore {
+export interface ViewPreferenceStore {
   /** Determines if the store has been hydrated from AsyncStorage. */
   _hasHydrated: boolean;
   /** Get a more accurate initial state. */
-  _init: (state: HomeViewPreferenceStore) => Promise<void>;
+  _init: (state: ViewPreferenceStore) => Promise<void>;
 
   artistLayout: LayoutOption;
 }
@@ -17,5 +17,5 @@ export interface HomeViewPreferenceStore {
 export const OmittedFields: string[] = [
   "_hasHydrated",
   "_init",
-] satisfies Array<keyof HomeViewPreferenceStore>;
+] satisfies Array<keyof ViewPreferenceStore>;
 //#endregion
